refactor: extract ArticleMetaDetail component
