+++ apps/web/app/articles/[slug]/_features/Article.tsx
@@ -1,27 +1,7 @@
-import { Time } from "@/app/_components";
-import type { ArticleMeta } from "@/app/_features";
-import { MarkdownContent, Tag, getArticleMeta } from "@/app/_features";
-import { format, rootJoin } from "@/app/_utils";
+import { ArticleMetaDetail } from "./ArticleMetaDetail";
 
-const ArticleMetaDetail = ({
-  meta: { publishedAt: _publishedAt, title, tags },
-}: { meta: ArticleMeta }) => {
-  const publishedAt = format(_publishedAt);
-
-  return (
-    <div>
-      <h2 className="text-2xl font-extrabold">{title}</h2>
-      <div className="my-6 flex justify-between">
-        <div className="flex gap-2">
-          {tags.map((tag) => (
-            <Tag key={tag} tag={tag} />
-          ))}
-        </div>
-        <Time dateTime={publishedAt}>{publishedAt}</Time>
-      </div>
-    </div>
-  );
-};
+import { MarkdownContent, getArticleMeta } from "@/app/_features";
+import { rootJoin } from "@/app/_utils";
 
 interface Props {
   slug: string;
